+++ extensions/stelow/modules/index.ts
@@ -12,7 +12,6 @@ export { JsonFileStore } from './file-store';
 export type {
   TaskStatus,
   TaskItem,
-  InboxItem
 } from './task';
 
 export { TASK_ICONS } from './task';